Took the class definitions out of document.ready
For this to work:
- the views have `el` specified as a CSS selector,
without $(). This is because the el property is
evaluated when the class definition is parsed, and
we don't have JQuery then
- also the template declaration was moved in the
`initialize` function

I changed the application initializer to a plain
JS object, as it does not need any functionality
from Backbone.View.

Also changed some variable names to less lazy ones.